fix: RedissonConfig 삭제 및 Redis 설정 통합
