+++ backend/src/main/java/com/app/backend/global/config/RedisConfig.java
@@ -8,6 +8,9 @@ import com.fasterxml.jackson.databind.jsontype.PolymorphicTypeValidator;
 import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import com.fasterxml.jackson.module.paramnames.ParameterNamesModule;
+import org.redisson.Redisson;
+import org.redisson.api.RedissonClient;
+import org.redisson.config.Config;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -22,9 +25,9 @@ import org.springframework.data.redis.serializer.StringRedisSerializer;
 @Configuration
 @EnableRedisRepositories
 public class RedisConfig {
-    @Value("${spring.data.redis.host}")
+    @Value("${spring.data.redis.host:localhost}")
     private String host;
-    @Value("${spring.data.redis.port}")
+    @Value("${spring.data.redis.port:6379}")
     private int    port;
     @Value("${spring.data.redis.password}")
     private String password;
@@ -44,8 +47,8 @@ public class RedisConfig {
         mapper.registerModule(new JavaTimeModule());
 
         PolymorphicTypeValidator ptv = BasicPolymorphicTypeValidator.builder()
-            .allowIfBaseType(Object.class)
-            .build();
+                                                                    .allowIfBaseType(Object.class)
+                                                                    .build();
         mapper.activateDefaultTyping(ptv, ObjectMapper.DefaultTyping.EVERYTHING, JsonTypeInfo.As.PROPERTY);
 
         RedisTemplate<String, Object> template = new RedisTemplate<>();
@@ -53,5 +56,20 @@ public class RedisConfig {
         template.setKeySerializer(new StringRedisSerializer());
         template.setValueSerializer(new GenericJackson2JsonRedisSerializer(mapper));
         return template;
+    }
+
+    @Bean
+    public RedissonClient redissonClient() {
+        Config config = new Config();
+        if (password != null && !password.isBlank())
+            config.useSingleServer()
+                  .setAddress("redis://%s:%d".formatted(host, port))
+                  .setPassword(password)
+                  .setConnectionPoolSize(64);
+        else
+            config.useSingleServer()
+                  .setAddress("redis://%s:%d".formatted(host, port))
+                  .setConnectionPoolSize(64);
+        return Redisson.create(config);
     }
 }
